+++ src/middlewared/middlewared/test/integration/assets/virt.py
@@ -16,7 +16,17 @@ def virt(pool_data: dict | None = None):
     try:
         yield virt_config
     finally:
-        virt_config = call('virt.global.update', {'pool': None, 'storage_pools': []}, job=True)
+        with contextlib.suppress(ValueError):
+            virt_config['storage_pools'].remove(pool_name)
+
+        virt_config = call(
+            'virt.global.update',
+            {
+                'pool': None,
+                'storage_pools': virt_config['storage_pools']
+            },
+            job=True
+        )
         assert virt_config['pool'] is None, virt_config
